pyproject.toml
[tool.poetry]
name = "ffxiv_gsheet_manage"
version = "0.1.0"
description = "A python tool for displaying and comparing character collectibles through Google Spreadsheets."

readme = "README.md"

authors = ["Aaron Declan hickey"]

[tool.poetry.dependencies]
python = "^3.8.0"
click = "8.1.3"
google-api-python-client = "2.72.0"
google-auth = "2.16.0"
PyYAML = "6.0"
XlsxWriter = "3.0.7"
